+++ lib/src/clixon_netconf_lib.c
@@ -1380,6 +1380,9 @@ netconf_module_load(clicon_handle h)
     if (clicon_option_bool(h, "CLICON_XML_CHANGELOG"))
 	if (yang_spec_parse_module(h, "clixon-xml-changelog", NULL, yspec)< 0)
 	    goto done;
+    /* Load restconf collection */
+    if (yang_spec_parse_module(h, "ietf-netconf-collection", NULL, yspec)< 0)
+	goto done;    
     retval = 0;
  done:
     return retval;